Move test setting data into class and change control loop structure
frequency, amplitude and control mode now are local variables in the SerialComm class

Changing the control modes is now much easier. One task, named Control-Loop-Task, loops indefinetly, starting off by waiting for a new item to be added to the test_setting_queue. Once a new item is added, it checks to make sure that it isn't a duplicate. If it isn't, it cancels the current Heater-Control-Task, waits for it to close, and then starts a new task up based on the self.CONTROL_OPTIONS dictionary.